+++ src/plugins/oSnap/components/TransactionBuilder/RawTransaction.vue
