+++ src/plugins/oSnap/components/TransactionBuilder/TokensModal.vue
