+++ src/plugins/oSnap/components/TransactionBuilder/TokensModalItem.vue
