+++ src/plugins/oSnap/components/TransactionBuilder/Transaction.vue
