+++ src/plugins/oSnap/components/TransactionBuilder/TransferFunds.vue
