+++ src/plugins/oSnap/components/TransactionBuilder/TransferNFT.vue
